+++ src/features/game/lib/constants.ts
@@ -370,7 +370,7 @@ export const INITIAL_EQUIPMENT: BumpkinParts = {
 
 export const INITIAL_BUMPKIN: Bumpkin = {
   equipped: INITIAL_EQUIPMENT as Equipped,
-  experience: 0,
+  experience: 10000,
 
   id: 1,
   skills: {},
@@ -670,7 +670,10 @@ export const INITIAL_FARM: GameState = {
       total: 10,
     },
   },
-  farmActivity: {},
+  farmActivity: {
+    "welcome Bonus Claimed": 1, // Skip welcome screen
+  },
+  username: "local",
   milestones: {},
   specialEvents: {
     history: {},
