+++ src/features/island/hud/components/settings-menu/GameOptions.tsx
@@ -63,7 +63,6 @@ import { ApiKey } from "./general-settings/ApiKey";
 import { ExperimentsSettings } from "./experiments-settings/ExperimentsSettings";
 import { EconomyEditorExperimentSettings } from "./experiments-settings/EconomyEditorExperimentSettings";
 import { InteriorExperimentSettings } from "./experiments-settings/InteriorExperimentSettings";
-import { DesignShowcaseSettings } from "./experiments-settings/DesignShowcaseSettings";
 import type { ContentComponentProps, SettingMenuId } from "./types";
 import { TwitterRewards } from "features/auth/components/Twitter/Twitter";
 import { TelegramBody } from "features/auth/components/Telegram/Telegram";
@@ -381,11 +380,6 @@ export const SETTING_MENUS: Record<SettingMenuId, SettingMenu> = {
     title: translate("gameOptions.experiments.interiors"),
     parent: "experiments",
     content: InteriorExperimentSettings,
-  },
-  designShowcase: {
-    title: translate("gameOptions.experiments.designShowcase"),
-    parent: "experiments",
-    content: DesignShowcaseSettings,
   },
 
   // Account
