+++ src/lib/i18n/dictionaries/dictionary.json
@@ -2207,7 +2207,6 @@
   "gameOptions.experiments.interiorsDescription": "Try the new interiors experience. When enabled, clicking your home takes you to the upgradeable interior instead of the classic home.",
   "gameOptions.experiments.interiorsToggle": "Enable new interiors",
   "gameOptions.experiments.interiorsDisclaimer": "Experimental feature only. Data and progress may be reset or lost without notice. By enabling this you accept these risks and proceed at your own risk.",
-  "gameOptions.experiments.designShowcase": "Design showcase",
   "designShowcase.heading": "Featured designs",
   "designShowcase.description": "A showcase of standout farm designs from the community.",
   "designShowcase.error": "Something went wrong loading the featured designs.",
